[#31] feat: Move customObject to shared type pkg
