+++ be/douren-backend/package.json
@@ -20,6 +20,7 @@
     "@hono/trpc-server": "^0.3.2",
     "@hono/zod-validator": "^0.2.2",
     "@pkg/database": "*",
+    "@pkg/type": "*",
     "@trpc/server": "^10.45.2",
     "@upstash/redis": "^1.34.0",
     "drizzle-orm": "^0.33.0",
