+++ pkg/type/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@pkg/type",
+  "type": "module",
+  "version": "0.0.1",
+  "scripts": {
+    "build": "tsc",
+    "build:pkg": "tsc"
+  },
+  "exports": {
+    "./model": {
+      "types": "./src/index.ts",
+      "default": "./dist/index.js"
+    }
+  },
+  "devDependencies": {
+    "@pkg/database": "*",
+    "@pkg/typescript-config": "*"
+  },
+  "dependencies": {
+    "drizzle-orm": "^0.33.0"
+  }
+}
