+++ pkg/type/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@pkg/typescript-config/pkg.json",
+  "compilerOptions": {
+    "outDir": "dist",
+    "rootDir": "src"
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "dist"]
+}
